requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Enhanced AI capabilities (FREE)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.0.0
transformers>=4.20.0

# Free LLM API options (choose one or more)
google-generativeai>=0.3.0    # Google Gemini
groq>=0.4.0                   # Groq API
cohere>=4.0.0                 # Cohere API  
huggingface_hub>=0.16.0       # HuggingFace

# Optional paid APIs
# openai>=1.0.0               # GPT-4 (if you get API key later)
# anthropic>=0.3.0            # Claude (if you get API key later)

# System utilities
requests>=2.25.0